+++ packages/static/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+
+* Remove `insert` option to `staticFiles`. `render` should return an HTML string.
+* Remove `redirects` option. `render` should throw for routes that shouldn't be rendered.
+
 ## 2.0.0-alpha.0
 
 * Bump to `v2` alpha since it uses `@curi/router` v2.